2 Teddies have to be compensated by chocolates from (2) And for each day a may be different. 563 $533116 55331760Given the maximum number of available Teddys M. andt
It's Tina's birthday today, and everyones bringing different gifts for her. You forgot to bring a gift for her. Se she became a bit sad.

955eer her up, you have pronised to give her Teddys for the next N days. But how many Teddys, that Tina will decide. She gave a list 'B' where she has mentioned the number of Teddys for each day from 1 to N, which she wants. 85331160 25

The problem is that there is a limited supply of Teddys in the cry. So, she decided to compensate for it with chocolates. But she has also set a rule for each day for chocolates also. For every missing 955331160 25

Teddy on day 1, you have to compensate it with CD]

number of chocolates per Teddy.

Let's say on day 2, you have to give 4 Teddies to Tina, but you gave only 2, then the remaining 2 Teddies have to be compensated by chocolates from (2) And for each day a may be different. 563 $533116 55331760

Given the maximum number of available Teddys M. and

the number of days N. You have to give the Teddies